requirements.txt
# === Django Core ===
Django==4.2
asgiref==3.7.2
sqlparse==0.4.4
pytz==2023.3.post1

# === Django Admin & UI ===
django-jazzmin==2.6.0

# === Django Extensions & Plugins ===
django-extensions==3.2.3
django-ckeditor-5==0.2.10
django-recaptcha==3.0.0
django-anymail==9.1
django-storages==1.12.3
django-redis==6.0.0
django-environ==0.12.0
django-debug-toolbar>=4.0

# === Database, Channels & Caching ===
dj-database-url==0.5.0
psycopg2-binary==2.9.9
redis==5.0.1

# === File Storage & Cloud Integration ===
boto3==1.20.26
pillow==11.2.1
whitenoise==5.2.0

# === Payments ===
stripe==7.0.0
shortuuid==1.0.11

# === API, Web, and Misc Helpers ===
requests==2.31.0
python-slugify==8.0.4

# === DevOps & Deployment ===
gunicorn==21.2.0
setuptools==75.8.0
packaging==23.2